sql_update_AccessRights

+++ sql/sql_update_AccessRights.sql
@@ -10,14 +10,18 @@ BEGIN
 		INSERT INTO mitarbeiter VALUES (NULL,p_User_ID,p_Level,p_Abteilung);
 	END IF;
 	-- Upgrade Mitarbeiter zu Administrator
-	IF ((@err=0) && (p_alt = "Mitarbeiter" || p_neu="Administrator") ) THEN
+	IF ((@err=0) && (p_alt = "Mitarbeiter") ) THEN
 		SELECT Mitarbeiter_ID FROM mitarbeiter WHERE User_ID = p_User_ID INTO @MID;
-		INSERT INTO administrator VALUES (NULL,@MID);
+		IF (p_neu = "Administrator") THEN
+			INSERT INTO administrator VALUES (NULL,@MID);
+		END IF;
+		UPDATE mitarbeiter SET Level = p_Level, Abteilung_ID = p_Abteilung WHERE Mitarbeiter_ID = @MID;
 	END IF;
 	-- Degradierung von Admin zu Mitarbeiter
-	IF (@err=0) && (p_alt = "Administrator") THEN
+	IF (@err=0) && (p_alt = "Administrator" && p_neu = "Mitarbeiter") THEN
 		SELECT Mitarbeiter_ID FROM mitarbeiter WHERE User_ID = p_User_ID INTO @MID;
 		DELETE FROM administrator WHERE Mitarbeiter_ID = @MID;
+		UPDATE mitarbeiter SET Level = p_Level, Abteilung_ID = p_Abteilung WHERE Mitarbeiter_ID = @MID;
 	END IF;
 	SET @ret=1;
 END //
